+++ src/components/Form/formChangelogData.ts
@@ -19,6 +19,7 @@ export const changelogData: ChangelogData[] = [
       "Removes `action`, `method`, and `onSubmit` props (as HTML equivalents may be used) and updates docs.",
       "Removes children validation warning.",
       "Updates the `id` prop to be optional and generates a random id if not provided.",
+      "Removes the explicit `classname` prop in favor of `rest`.",
     ],
   },
   {