+++ src/components/RadioGroup/radioGroupChangelogData.tsx
@@ -18,6 +18,8 @@ export const changelogData: ChangelogData[] = [
       "Removes explicit `className` prop as interface can be extended to include Chakra prop or HTML attribute types.",
       "Replaces `Children.map` logic with `RadioGroupContext`.",
       "Updates the `id` prop to be optional and generates a random id if not provided.",
+      "Removes the explicit `classname` prop in favor of `rest`.",
+      "Renames `radioGroupStack` to `ds-radioGroup-stack` to align with the new naming convention.",
     ],
   },
   {