+++ src/components/SearchBar/searchBarChangelogData.ts
@@ -20,6 +20,7 @@ export const changelogData: ChangelogData[] = [
       "Removes explicit `className` prop as interface can be extended to include Chakra prop or HTML attribute types.",
       "Removes `action`, `method`, and `onSubmit` props (as HTML equivalents may be used) and updates docs.",
       "Updates the `id` prop to be optional and generates a random id if not provided.",
+      "Renames internal classnames `textInput` and `searchButton` to `ds-searchBar-textInput` and `ds-searchBar-button` to align with the new naming convention.",
     ],
   },
   {