+++ src/components/SkeletonLoader/skeletonLoaderChangelogData.ts
@@ -17,6 +17,7 @@ export const changelogData: ChangelogData[] = [
     notes: [
       "Removes explicit `className` prop as interface can be extended to include Chakra prop or HTML attribute types.",
       "Updated the internal use of ids to reduce `undefined` ids in the DOM.",
+      "Removes internal 'row' or 'column' classname.",
     ],
   },
   {